Finalizado soma de pontos, necessario aplicar em records
