+++ modules/checkForWin.js
@@ -1,0 +1,25 @@
+
+const linesPlayed = document.querySelectorAll(".line-played");
+const winningCombinations = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8],
+    [2, 4, 6]
+];
+
+const checkForWin = (currentPlayer) => {
+    return winningCombinations.some(combination => {
+        return combination.every(index => {
+            return linesPlayed[index].classList.contains(currentPlayer);
+        });
+    });
+};
+
+export {
+    checkForWin,
+    linesPlayed
+}
