+++ modules/createCells.js
@@ -1,0 +1,6 @@
+
+
+export default createDataCells = (cell) => {
+    const td = document.createElement('td');
+    td.innerText = `${cell}`;
+}
