+++ modules/placeMark.js
@@ -1,0 +1,5 @@
+
+
+export default placeMark = (line, classToAdd) => {
+    line.classList.add(classToAdd);
+}
