+++ modules/swapTurn.js
@@ -1,0 +1,13 @@
+import setBoardHoverClass from "./setBoardHoverClass";
+
+let isCircleTurn;
+
+const swapTurn = () => {
+    isCircleTurn = !isCircleTurn;
+    setBoardHoverClass();
+}
+
+export {
+    isCircleTurn,
+    swapTurn
+}
