+++ records.html
@@ -14,15 +14,13 @@
     <table class="table">
         <thead>
             <tr>
-                <th id="winners" class="table-th">Vencedores</th>
-                <th id="draws" class="table-th">Velhas</th>
+                <th class="table-th">Vencedores</th>
+                <th class="table-th">Velhas</th>
             </tr>
         </thead>
         <tbody class="table-tbody">
-            <tr>
-                <td></td>
-                <td></td>
-            </tr>
+            <tr id="winners"></tr>
+            <tr id="draws"></tr>
         </tbody>
     </table>
     <div>
