+++ records.js
@@ -1,0 +1,5 @@
+import showRecords from './modules/showRecords.js';
+
+const recordsButton = document.querySelector("#buttonRecords");
+
+recordsButton.addEventListener('click', showRecords());
